Fix: Include all CO2 calculation parameters into trace and provenance
Feat: Partial fix of rounding errors by using BigDecimal instead of Double

Test: Adjusted number rounding in tests to fit new BigDecimal format

Test: Adjusted output file checks

Chore: Optimized imports

+++ src/main/nextflow/co2footprint/DataContainers/CIDataMatrix.groovy
@@ -82,7 +82,7 @@ class CIDataMatrix extends DataMatrix {
      */
     CIMatch findCiInMatrix(String targetZone, String fallbackZone=null) throws IllegalStateException {
         if (rowIndex.containsKey(targetZone)){
-            Number ci = get(targetZone, this.ciColumn) as Double
+            Number ci = get(targetZone, this.ciColumn) as BigDecimal
             return new CIMatch(targetZone, ci)
         }
         else if (fallbackZone != null) {
